backend-fastapi-pymongo/backend/requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.1.0
bcrypt==3.2.2
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.0.4
fastapi==0.88.0
flit_core==3.8.0
gmpy2==2.1.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.2
httpx==0.23.1
hyperframe==6.0.1
idna==3.4
iniconfig==1.1.1
packaging==22.0
passlib==1.7.4
pip==22.3.1
pluggy==1.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pymongo==4.3.3
pyOpenSSL==22.1.0
PySocks==1.7.1
pytest==7.2.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
setuptools==65.5.1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
wheel==0.38.4
win-inet-pton==1.1.0
